% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\name{download_cities_with_population_5000}
\alias{download_cities_with_population_5000}
\title{Download cities with population 5000}
\usage{
download_cities_with_population_5000()
}
\description{
Download cities with population 5000 information from geonames.
}
\details{
It returns a tibble with all the data
}
\examples{
download_cities_with_population_5000()

}
\seealso{
<https://github.com/jrosell/rreversegeocoder/blob/main/R/cities.R>
}
