% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_geocoder.R
\name{reverse_geocoder_us_cities_with_population_5000}
\alias{reverse_geocoder_us_cities_with_population_5000}
\title{Reverse geocoder}
\usage{
reverse_geocoder_us_cities_with_population_5000(df)
}
\arguments{
\item{df}{A data.frame or tibble with the first two columns latitude and longitude}
}
\description{
Given latitude and longitude, find the nearest city in US with population 5000.
}
\details{
It returns the nearest geonameid from geonames.
}
\examples{
reverse_geocoder_us_cities_with_population_5000(data.frame(34.6, -120))

}
\seealso{
<https://github.com/jrosell/rreversegeocoder/blob/main/R/reverse_geocoder.R>
}
